changed TransferService logic, so it handles fraction weights as well. Changed necessary  int attributes to double, the project should be done.

+++ src/main/java/ge/zerobyte/cheapesttransferroute/api/controller/TransferController.java
@@ -46,7 +46,7 @@ public class TransferController {
 
         // Extract the necessary information from the request.
         List<Transfer> availableTransfers = requestData.getAvailableTransfers();
-        int maxWeight = requestData.getMaxWeight();
+        double maxWeight = requestData.getMaxWeight();
 
         // Pass it to the service and get a result.
         List<Transfer> selectedTransfers = transferService.getCheapestTransferRoute(availableTransfers, maxWeight);
